Move MatchSearchQuery to its own package
